Added GCE node builder, tweak to run environment
The GCE node builder is very similar to AWS, just a different
spack archive is used.

The run script needs to explicitly ~/.bashrc if run in an automated
way (local_setup.sh changes .bashrc, but those changes are not
available in the parent shell if run_aws_c5n9x.sh is run immediately
afterward).