Temporarily allow --no-check-signature in build
I am part way through the final WRF build but running
out of memory. This change allows me to migrate to a
bigger node, manually clone spack, add the buildcache,
and reuse most of the build artefacts on a bigger node.
Remove this option later so that the primary build
always checks signatures on download.