Add SPACK_ROOT sourcing lines to ~/.bashrc instead of /etc/bashrc
I originally used /etc/bashrc to make this change persistent if
and image snapshot were to be taken. But it won't work if this
cluster is used as a shared cluster. Since we are taking a new
approach of build-on-the-fly with a buildcache, make this change
just local to the user instead of a persistent system-wide change.